meta: refactor data generation to server-side only
