+++ .eslintignore
@@ -6,10 +6,6 @@ node_modules
 .swc
 build
 
-# Public Files
-public/node-releases-data.json
-public/blog-posts-data.json
-
 # We don't want to lint/prettify the Coverage Results
 coverage
 junit.xml
