+++ .husky/pre-commit
@@ -1,9 +1,5 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-# if the generated files got tracked to this commit we revert them
-git reset public/node-releases-data.json
-git reset public/blog-posts-data.json
-
 # lint and format staged files
 npx lint-staged
