+++ .prettierignore
@@ -9,8 +9,6 @@ build
 
 # Next.js Generated Files
 public/static/documents
-public/node-releases-data.json
-public/blog-posts-data.json
 
 # Jest
 coverage
