+++ layouts/BaseLayout.tsx
@@ -4,10 +4,13 @@ import type { FC, PropsWithChildren } from 'react';
 
 import Footer from '@/components/Footer';
 import Header from '@/components/Header';
+import TopNavigation from '@/components/TopNavigation';
 
 const BaseLayout: FC<PropsWithChildren> = ({ children }) => (
   <>
-    <Header />
+    <Header>
+      <TopNavigation />
+    </Header>
     <main id="main">{children}</main>
     <Footer />
   </>
