+++ next.data.mjs
@@ -4,8 +4,11 @@ import generateBlogPostsData from './next-data/generateBlogPostsData.mjs';
 import generateNodeReleasesJson from './next-data/generateNodeReleasesJson.mjs';
 import generateWebsiteFeeds from './next-data/generateWebsiteFeeds.mjs';
 
-export {
-  generateWebsiteFeeds,
-  generateBlogPostsData,
-  generateNodeReleasesJson,
-};
+/** @type {import('./types').BlogData} */
+export const blogData = await generateBlogPostsData();
+
+/** @type {import('./types').NodeRelease[]} */
+export const releaseData = await generateNodeReleasesJson();
+
+/** @type {Map<string, import('feed').Feed} */
+export const websiteFeeds = await generateWebsiteFeeds(blogData);
