--- public/blog-posts-data.json
@@ -1,1 +1,0 @@
-{ "pagination": [], "categories": [], "posts": [] }
