--- public/node-releases-data.json
@@ -1,1 +1,0 @@
-[]
